updateShape function updated and the realted codes were refactered - change the call of the function from ref.listen to the onChange parameter of Dropdown Button
